+++ app/src/main/java/com/example/gafitorsatpam/MainActivity.kt
@@ -15,8 +15,8 @@ import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import com.example.gafitorsatpam.main.NotificationMessage
 import com.example.gafitorsatpam.auth.LoginScreen
-import com.example.gafitorsatpam.ui.fe.LaporanSatpam.LaporKehilanganScreen
-import com.example.gafitorsatpam.ui.fe.LaporanSatpam.DetailLapScreen
+import com.example.gafitorsatpam.ui.fe.laporanSatpam.LaporKehilanganScreen
+import com.example.gafitorsatpam.ui.fe.laporanSatpam.DetailLapScreen
 import com.example.gafitorsatpam.ui.fe.ParkirDataScreen
 import com.example.gafitorsatpam.ui.fe.parkir.LaporParkirScreen
 import com.example.gafitorsatpam.ui.theme.GafitorSatpamTheme
