+++ app/src/main/java/com/example/gafitorsatpam/ui/fe/laporanSatpam/LaporKehilanganScreen.kt
@@ -1,47 +1,44 @@
-package com.example.gafitorsatpam.ui.fe.LaporanSatpam
+package com.example.gafitorsatpam.ui.fe.laporanSatpam
 
 import android.annotation.SuppressLint
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Scaffold
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.example.gafitorsatpam.component.TopBarAtas
-import com.example.gafitorsatpam.component.laporanComp.DetailLaporan
+import androidx.navigation.NavController
+import com.example.gafitorsatpam.GafitoViewModel
+import com.example.gafitorsatpam.component.BottomBar
+import com.example.gafitorsatpam.component.TopBarMenu
 import com.example.gafitorsatpam.component.laporanComp.FormLaporan
+import com.example.gafitorsatpam.model.BottomBarItem
 import com.example.gafitorsatpam.ui.theme.GafitorSatpamTheme
 
 @OptIn(ExperimentalMaterial3Api::class)
 @SuppressLint("UnusedMaterial3ScaffoldPaddingParameter")
 @Composable
-fun DetailLapScreen() {
+fun LaporKehilanganScreen(navController: NavController, vm: GafitoViewModel, encodedUri: String) {
     Scaffold(
-        topBar = {TopBarAtas("Detail Laporan")}
+        topBar = { TopBarMenu(screen = "Lapor Kehilangan") },
+        bottomBar = {BottomBar(
+            selectedItem = BottomBarItem.REPORT,
+            navController = navController)}
     ) {
-            paddingValues ->
         Column(
-            verticalArrangement = Arrangement.Center,
-            horizontalAlignment = Alignment.CenterHorizontally,
             modifier = Modifier
-                .padding(paddingValues)
-
+                .padding(top = 64.dp)
         ) {
-//        your code compose here
-            DetailLaporan()
-
+            FormLaporan(navController, vm, encodedUri)
         }
     }
 }
 
-@Preview(showBackground = true)
+@Preview (showBackground = true)
 @Composable
-fun DetailLaporanPrev() {
+fun LaporHilangPreview() {
     GafitorSatpamTheme {
-        DetailLapScreen()
     }
 }
