+++ app/src/main/java/com/example/gafitorsatpam/ui/fe/laporanSatpam/ListLaporanScreen.kt
@@ -1,9 +1,8 @@
-package com.example.gafitorsatpam.ui.fe.LaporanSatpam
+package com.example.gafitorsatpam.ui.fe.laporanSatpam
 
 import android.annotation.SuppressLint
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Scaffold
@@ -11,9 +10,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
 import com.example.gafitorsatpam.component.TopBarAtas
-import com.example.gafitorsatpam.component.laporanComp.FormLaporan
 import com.example.gafitorsatpam.component.laporanComp.ListLaporanView
 import com.example.gafitorsatpam.ui.theme.GafitorSatpamTheme
